# DZ5/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления шахматными играми",
        "title": "Шахматы API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/game": {
            "post": {
                "description": "Создаёт игру с заданным размером доски и именами игроков",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Создать новую игру",
                "parameters": [
                    {
                        "description": "Параметры игры",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateGameRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/game/{id}": {
            "get": {
                "description": "Возвращает полное состояние игры, включая доску",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Получить игру по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GameResponse"
                        }
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Выполняет действие над игрой (например, сдача)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Обновить игру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Действие",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет игру по ID из памяти и файла",
                "tags": [
                    "game"
                ],
                "summary": "Удалить игру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Игра удалена"
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/game/{id}/auto": {
            "post": {
                "description": "Сервер генерирует и выполняет указанное количество случайных ходов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Автоматические ходы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Количество ходов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AutoPlayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/game/{id}/move": {
            "post": {
                "description": "Выполняет ход фигуры в указанной игре",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Сделать ход",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ход",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.MoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка хода",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/games": {
            "get": {
                "description": "Возвращает массив всех игр с основной информацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "Получить список всех игр",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.GameInfo"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.AutoPlayRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "handlers.CreateGameRequest": {
            "type": "object",
            "properties": {
                "first_player": {
                    "type": "string",
                    "example": "Alice"
                },
                "second_player": {
                    "type": "string",
                    "example": "Bob"
                },
                "size": {
                    "type": "integer",
                    "example": 8
                }
            }
        },
        "handlers.GameInfo": {
            "type": "object",
            "properties": {
                "current_turn": {
                    "type": "string"
                },
                "first_player": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "second_player": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.GameResponse": {
            "type": "object",
            "properties": {
                "board": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "current_player": {
                    "type": "string"
                },
                "first_player": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moves_made": {
                    "type": "integer"
                },
                "second_player": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.MoveRequest": {
            "type": "object",
            "properties": {
                "from_x": {
                    "type": "integer",
                    "example": 4
                },
                "from_y": {
                    "type": "integer",
                    "example": 6
                },
                "to_x": {
                    "type": "integer",
                    "example": 4
                },
                "to_y": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
        "handlers.UpdateGameRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "surrender"
                }
            }
        }
    }
}